pcb: A few tweaks for next version
Changed the footprint of the 10nF's to 0603, as I've discovered I have a
bout 1Kg of them.
Also changed pinout! Silly me, I used A6 and A7 pins on the arduidiot,
but these are ADC only, no digital logic. So luckily I had hooked up two
'spare' pins, so I switched to these instead.

Also, the rest of the PCB has beem tested, and the switch is defaulting
to the RX mode when EN pin is low (as expected), so the pulldown works
fine, and the AVR can keep it's antenna switch pin as open drain.

Signed-off-by: Michel Pollet <buserror@gmail.com>